+++ tests/test_inflector_ca.py
@@ -14,14 +14,14 @@ def test_all_verbs_have_templates():
     verbs = cg.get_verbs()
     assert len(verbs) == 8616
     template_names = cg.get_template_names()
-    assert len(template_names) == 48
+    assert len(template_names) == 49
     missing_verbs = []
     missing_templates = set()
     for verb in verbs:
         if verb.template not in template_names:
             missing_templates.add(verb.template)
             missing_verbs.append(verb)
-    # assert len(missing_templates) == 0
+    assert len(missing_templates) == 0
     # assert len(missing_verbs) == 0
 
 
@@ -1859,6 +1859,21 @@ def test_find_verb_by_infinitive():
                 "nosaltres valguérem",
                 "vosaltres valguéreu",
                 "elles valgueren",
+            ],
+        ),
+        (
+            "doldre",
+            "indicatiu",
+            "present",
+            True,
+            "f",
+            [
+                "jo dolc",
+                "tu dols",
+                "ella dol",
+                "nosaltres dolem",
+                "vosaltres doleu",
+                "elles dolen",
             ],
         ),
     ],