+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 - 1.9.4 [23 December 2023]
   - Improved Catalan Support
-    - Added more missing templates (6+)
-    - TODO: Still missing 30 templates for 184 out of 8616 verbs
+    - Added more missing templates (7+)
+    - TODO: Still missing 29 templates for 179 out of 8616 verbs
     - See `test_inflector_ca.test_all_verbs_have_templates`
 
 - 1.9.3 [22 December 2023]